Move presenters and related helpers to main/
