+++ app/main/helpers/search_helpers.py
