+++ tests/main/presenters/test_search_summary.py
@@ -4,9 +4,9 @@ import json
 from flask import Markup
 from mock import Mock
 from werkzeug.datastructures import MultiDict
-from app.presenters.search_presenters import filters_for_lot
-from app.presenters.search_results import SearchResults
-from app.presenters.search_summary import SearchSummary, \
+from app.main.presenters.search_presenters import filters_for_lot
+from app.main.presenters.search_results import SearchResults
+from app.main.presenters.search_summary import SearchSummary, \
     SummaryRules, SummaryFragment
 from app import content_loader
 
@@ -19,7 +19,7 @@ filter_groups = filters_for_lot(
 
 def _get_fixture_data():
     test_root = os.path.abspath(
-        os.path.join(os.path.dirname(__file__), "..")
+        os.path.join(os.path.dirname(__file__), "../..")
     )
     fixture_path = os.path.join(
         test_root, 'fixtures', 'search_results_fixture.json'
@@ -30,7 +30,7 @@ def _get_fixture_data():
 
 def _get_fixture_multiple_pages_data():
     test_root = os.path.abspath(
-        os.path.join(os.path.dirname(__file__), "..")
+        os.path.join(os.path.dirname(__file__), "../..")
     )
     fixture_path = os.path.join(
         test_root, 'fixtures', 'search_results_multiple_pages_fixture.json'
